pyproject.toml
[tool.poetry]
name = "jetscape_stat"
version = "0.9.0"
description = "JETSCAPE Bayesian Parameter Estimation"
authors = [
    "Jake Coleman",
    "Jonah Bernhard <jonah.bernhard@gmail.com>",
    "Yi Chen <yi.chen@cern.ch>",
    "Weiyao Ke <wk42@phy.duke.edu>",
    "James Mulligan <james.mulligan@berkeley.edu>",
    "Raymond Ehlers <raymond.ehlers@cern.ch>",
]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.1"
emcee = "<=2.2.1"
h5py = "^2.10.0"
numpy = "^1.19.0"
PyYAML = "^5.3.1"
joblib = "^0.15.1"
scikit-learn = "^0.23.1"
scipy = "^1.5.0"
pandas = "^1.0.5"
pathlib = "^1.0.1"
hsluv = "^5.0.0"
matplotlib = "^3.2.2"
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
